Initial take on the human output
