+++ stats/stats.go
@@ -9,12 +9,6 @@
 package stats
 
 import (
-	"encoding/json"
-	"fmt"
-	"sort"
-
-	"github.com/abhiyerra/gojsonexplode"
-
 	"github.com/acksin/procfs"
 	"github.com/acksin/strum/io"
 	"github.com/acksin/strum/memory"
@@ -65,47 +59,6 @@ func (n *Stats) processes() procfs.Procs {
 	}
 
 	return procs
-}
-
-// JSON returns JSON string of Stats
-func (n *Stats) JSON() string {
-	js, err := json.MarshalIndent(n, "", "  ")
-	if err != nil {
-		return ""
-	}
-
-	return string(js)
-}
-
-// Flat returns a flattened results.
-func (n *Stats) Flat() string {
-	o, err := gojsonexplode.Explodejsonstr(n.JSON(), ".")
-	if err != nil {
-		fmt.Println(err)
-		return ""
-	}
-
-	var out map[string]interface{}
-
-	err = json.Unmarshal([]byte(o), &out)
-	if err != nil {
-		fmt.Println(err)
-		return ""
-	}
-
-	var o2 string
-	var keys []string
-
-	for k := range out {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-
-	for _, k := range keys {
-		o2 += fmt.Sprintf("%s = %v\n", k, out[k])
-	}
-
-	return o2
 }
 
 func (n *Stats) containsPid(pids []int, proc procfs.Proc) bool {
